requirements.txt
streamlit==1.39.0
openai==1.54.4
pydantic==2.9.2
pyyaml==6.0.2
plotly==5.24.1
pandas==2.2.3
python-dotenv==1.0.1
rich==13.9.4
tenacity==9.0.0
